Rescue uncommitted DGX/cross-session WIP from working tree
These changes were sitting uncommitted in the working tree at session
start. Surfacing them here so they're not lost; not authored in this
session — flag for review.

  jaxoccoli/glm.py:        jnp.linalg.inv(XtX) → cholesky_invert(XtX + 1e-6 I)
                           — better numerical stability, particularly
                             relevant to the AR(1) parity work
  jaxoccoli/realtime.py:   companion change
  pyproject.toml:          1 line addition
  tests/test_geometry.py:  refactor

  tests/test_riemannian_ml.py and tests/test_sklearn_compliance.py:
                           DELETED — both were skip-only on this Mac
                           (sklearn not in venv, riemannian_ml has its
                           own dependency stack)

  scripts/preproc_bakeoff.py:        new benchmark over Variant G /
                                     streaming Kalman / Riemannian
  scripts/task_2_1_top_candidates.py: top-candidate rerun script

Smoke-tested: jaxoccoli imports + GLM constructs cleanly with the new
Cholesky path; both new scripts parse.

+++ jaxoccoli/realtime.py
@@ -109,10 +109,10 @@ def build_lss_design_matrix(events_onsets: np.ndarray, tr: float, n_trs: int,
 
 @partial(jax.jit, static_argnames=())
 def _variant_g_forward_jit(X_pad, Y_pad, n_eff,
-                           pp_scalar=jnp.float32(0.01),
-                           rho_prior_mean=jnp.float32(0.5),
-                           rho_prior_var=jnp.float32(0.09),
-                           a0=jnp.float32(0.01), b0=jnp.float32(0.01)):
+                           pp_scalar=0.01,
+                           rho_prior_mean=0.5,
+                           rho_prior_var=0.09,
+                           a0=0.01, b0=0.01):
     """AR(1)-prewhitened conjugate Gaussian GLM, padded inputs, vmapped voxels.
 
     Args:
